+++ lib/presentation/common/components/error_view.dart
@@ -1,3 +1,7 @@
+// Copyright 2022 susatthi All rights reserved.
+// Use of this source code is governed by a MIT license that can be
+// found in the LICENSE file.
+
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
