+++ lib/presentation/repo/components/order_toggle_button.dart
@@ -8,7 +8,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import '../../../domain/entities/repo_search_repos_order.dart';
 import '../../../localizations/strings.g.dart';
 import '../../../utils/logger.dart';
-import 'list_view_notifier.dart';
+import 'list_view_state.dart';
 import 'search_repos_order.dart';
 
 /// リポジトリ検索用オーダー値変更ボタン
