+++ test/presentation/repo/pages/view_page_test.dart
@@ -5,8 +5,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/presentation/repo/components/detail_view.dart';
 import 'package:github_search/presentation/repo/components/full_name_text.dart';
-import 'package:github_search/presentation/repo/components/selected_repo.dart';
 import 'package:github_search/presentation/repo/pages/view_page.dart';
+import 'package:github_search/presentation/repo/state/selected_repo.dart';
 
 import '../../../test_utils/locale.dart';
 import '../../../test_utils/mocks.dart';
