+++ test/presentation/repo/state/selected_repo_test.dart
@@ -8,8 +8,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/config/router.dart';
 import 'package:github_search/infrastructure/github/json_object/repo/repo.dart';
 import 'package:github_search/infrastructure/github/repo_repository.dart';
-import 'package:github_search/presentation/repo/components/selected_repo.dart';
 import 'package:github_search/presentation/repo/pages/view_page.dart';
+import 'package:github_search/presentation/repo/state/selected_repo.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../test_utils/locale.dart';
